[adding-payment-api-endpoint] Splitting up the PayPal controller into two controllers to increase readability
